heronPipeline/src/functions/createRunBaseConfig/requirements.txt

pandas
numpy
boto